Add initial minimal working map input file
This is a first-cut implementation of the railway network for Hannover.
Note that this is *not* intended to reflect the real-world situation just yet.
I've chosen to use station names here which make the initial validation tests
pass and which vaguely reflect the nature of the network itself.  Both
stations do exist on Linie 1, however are separated by several other stations
in reality.  Since the validation tests pass, we know that things are wired
up to the Map::Tube framework properly.